requirements.txt
aiosqlite>=0.20.0
anyio>=4.4.0
asyncpg>=0.29.0
fastapi>=0.115.0
fpdf2>=2.8.0
httpx>=0.27.0
langchain-core>=0.3.0
langfuse>=2.50.0
langgraph>=0.2.20
openai>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
uvicorn[standard]>=0.30.0
